updates to location of FAQ content files
