+++ react-app/src/features/faq/Faq.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { Navigate, useParams } from "react-router";
 
 import {
@@ -13,10 +13,20 @@ import {
 import { useFeatureFlag } from "@/hooks/useFeatureFlag";
 import { sendGAEvent } from "@/utils/ReactGA/SendGAEvent";
 
-import { helpDeskContact, oneMACFAQContent } from "./content/oneMACFAQContent";
+import { oneMACFAQContent } from "./faqs";
+import { helpDeskContact } from "./faqs/utils";
 
 export const Faq = () => {
   const isFAQHidden = useFeatureFlag("TOGGLE_FAQ");
+
+  // Get the flag value for hiding the MMDL banner.
+  const isBannerHidden = useFeatureFlag("UAT_HIDE_MMDL_BANNER");
+  const isUpdateNewLabel = useFeatureFlag("UPGRADE_NEW_LABEL");
+  const isChipSpaDetailsEnabled = useFeatureFlag("CHIP_SPA_DETAILS");
+
+  const faqContent = useMemo(() => {
+    return oneMACFAQContent({ isChipSpaDetailsEnabled, isBannerHidden });
+  }, [isChipSpaDetailsEnabled, isBannerHidden]);
 
   const { id } = useParams<{ id: string }>();
 
@@ -24,11 +34,11 @@ export const Faq = () => {
 
   const expandAll = useCallback(() => {
     const allIds = [];
-    oneMACFAQContent.forEach(({ qanda }) => {
+    faqContent.forEach(({ qanda }) => {
       qanda.forEach(({ anchorText }) => allIds.push(anchorText));
     });
     setOpenItems(allIds);
-  }, [setOpenItems]);
+  }, [setOpenItems, faqContent]);
 
   useEffect(() => {
     if (id) {
@@ -42,255 +52,6 @@ export const Faq = () => {
       }
     }
   }, [id]);
-
-  // Get the flag value for hiding the MMDL banner.
-  const isBannerHidden = useFeatureFlag("UAT_HIDE_MMDL_BANNER");
-  const isUpdateNewLabel = useFeatureFlag("UPGRADE_NEW_LABEL");
-  const isChipSpaDetailsEnabled = useFeatureFlag("CHIP_SPA_DETAILS");
-
-  const chipSpaFAQ = isChipSpaDetailsEnabled
-    ? {
-        anchorText: "chip-spa-attachments",
-        question: "What are the attachments for a CHIP SPA?",
-        label: null,
-        labelColor: null,
-        answerJSX: (
-          <>
-            <p className="font-bold">CHIP SPA Attachment Types:</p>
-            <p className="ml-12">Note: “*” indicates a required attachment.</p>
-            <table className="faq-table ml-12 border-collapse border border-gray-300 w-full ">
-              <tbody>
-                <tr>
-                  <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
-                  <th className="border border-gray-300 px-4 py-2">Description</th>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Current State Plan*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Current version of the CHIP state plan that details how the State operates its
-                    CHIP program
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
-                    Amended State Plan Language*
-                  </td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Track changes to <span className="underline">only</span> the currently approved
-                    CHIP state plan pages that the State is proposing to amend
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Cover Letter*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Cover letter to CMS with an authorized signature that outlines the purpose of
-                    the CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Updated 1-year budget if applicable of the State's planned expenditures if the
-                    CHIP SPA submission has a significant impact on the approved budget
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Public Notice</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Process used by the State if applicable to accomplish involvement of the public
-                    that occurred specifically for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Consultation process with Indian Tribes if applicable that occurred specifically
-                    for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Other</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Other supporting document(s) needed to process the CHIP SPA submission
-                  </td>
-                </tr>
-              </tbody>
-            </table>
-
-            <p className="font-bold mt-12">CHIP Eligibility SPA Attachment Types:</p>
-            <p className="ml-12">Note: “*” indicates a required attachment.</p>
-            <table className="faq-eligibility-table ml-12 border-collapse border border-gray-300 w-full ">
-              <tbody>
-                <tr>
-                  <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
-                  <th className="border border-gray-300 px-4 py-2">Description</th>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
-                    CHIP Eligibility Template*
-                  </td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Amendment to the eligibility section of the CHIP state plan using a PDF template
-                    repository, including statutory and regulatory background, required information,
-                    and minimum review criteria
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Cover Letter*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Cover letter to CMS with an authorized signature that outlines the purpose of
-                    the CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Current State Plan</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Current version of the CHIP state plan that details how the State operates its
-                    CHIP program
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
-                    Amended State Plan Language
-                  </td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Track changes to <span className="underline">only</span> the currently approved
-                    CHIP state plan pages that the State is proposing to amend
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Updated 1-year budget if applicable of the State's planned expenditures if the
-                    CHIP SPA submission has a significant impact on the approved budget
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Public Notice</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Process used by the State if applicable to accomplish involvement of the public
-                    that occurred specifically for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Consultation process with Indian Tribes if applicable that occurred specifically
-                    for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Other</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Other supporting document(s) needed to process the CHIP Eligibility SPA
-                    submission
-                  </td>
-                </tr>
-              </tbody>
-            </table>
-          </>
-        ),
-      }
-    : {
-        anchorText: "chip-spa-attachments",
-        question: "What are the attachments for a CHIP SPA?",
-        answerJSX: (
-          <>
-            <p>Note: “*” indicates a required attachment.</p>
-            <table className="faq-table border-collapse border border-gray-300 w-full ">
-              <tbody>
-                <tr>
-                  <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
-                  <th className="border border-gray-300 px-4 py-2">Description</th>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Current State Plan*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Current version of the CHIP state plan that details how the State operates its
-                    CHIP program
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Amended State Plan Language*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Redline version of proposed changes to the existing CHIP state plan pages. State
-                    to provide a redline version and a clean version of the CHIP state plan pages
-                    being amended.
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Cover Letter*</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Cover letter to CMS with an authorized signature that outlines the purpose of
-                    the CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Updated 1-year budget if applicable of the State's planned expenditures if the
-                    CHIP SPA submission has a significant impact on the approved budget
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Public Notice</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Process used by the State if applicable to accomplish involvement of the public
-                    that occurred specifically for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Consultation process with Indian Tribes if applicable that occurred specifically
-                    for this CHIP SPA submission
-                  </td>
-                </tr>
-                <tr>
-                  <td className="border border-gray-300 px-4 py-2">Other</td>
-                  <td className="border border-gray-300 px-4 py-2">
-                    Other document(s) needed to process the CHIP SPA submission
-                  </td>
-                </tr>
-              </tbody>
-            </table>
-          </>
-        ),
-      };
-
-  const anchorsToHide = [
-    "spa-admendments",
-    "abp-spa-templates",
-    "abp-implementation-guides-spa",
-    "mpc-spa-templates",
-    "mpc-spa-implementation-guides",
-    "chip-spa-templates",
-    "chip-spa-implentation-guides",
-  ];
-
-  const filteredFAQContent = oneMACFAQContent.map((section) => {
-    if (section.sectionTitle === "State Plan Amendments (SPAs)") {
-      const filteredQanda = isBannerHidden
-        ? section.qanda.filter((qa) => !anchorsToHide.includes(qa.anchorText))
-        : section.qanda;
-
-      const existingQanda = filteredQanda.filter((qa) => qa.anchorText !== "chip-spa-attachments");
-
-      // Find the correct index for chip-spa-attachments to appear before RAI question
-      const raiIndex = existingQanda.findIndex(
-        (qa) => qa.anchorText === "chip-spa-rai-attachments",
-      );
-
-      const updatedQanda = [
-        ...existingQanda.slice(0, raiIndex),
-        chipSpaFAQ,
-        ...existingQanda.slice(raiIndex),
-      ];
-
-      return { ...section, qanda: updatedQanda };
-    }
-    return section;
-  });
 
   if (isFAQHidden) {
     return <Navigate to="/" replace />;
@@ -321,7 +82,7 @@ export const Faq = () => {
 
             {/* FAQ */}
             <Accordion type="multiple" value={openItems} onValueChange={setOpenItems}>
-              {filteredFAQContent.map(({ sectionTitle, qanda }) => (
+              {faqContent.map(({ sectionTitle, qanda }) => (
                 <article key={sectionTitle} className="mb-8">
                   <h2 className="text-2xl mb-4 text-primary">{sectionTitle}</h2>
                   {qanda.map(({ anchorText, answerJSX, question, label, labelColor }) => (
