+++ react-app/src/features/faq/content/abpGuides.ts
@@ -1,4 +1,4 @@
-import { Template } from "./chpRenderSection";
+import { Template } from "../faqs/utils";
 
 export const ABP_GUIDES: Template[] = [
   {
